+++ lib/phobos/errors.rb
@@ -3,4 +3,5 @@
 module Phobos
   class Error < StandardError; end
   class AbortError < Error; end
+  class InvalidHandlerError < Error; end
 end